requirements.txt
langchain==0.2.2
langchain_community==0.2.3
langchain_openai==0.1.8
langchain_text_splitters==0.2.4
pdfplumber==0.11.4
python-dotenv==1.0.1
streamlit==1.38.0
chromadb==0.5.5
wrapt==1.16.0
opentelemetry-instrumentation==0.48b0
SQLAlchemy==2.0.34
